requirements.txt
streamlit
pandas~=1.3.5
tensorflow~=2.16.1
keras~=3.1.0
requests~=2.31.0
pillow~=10.2.0
scikit-multilearn~=0.2.0
tqdm~=4.66.2
numpy~=1.26.4
matplotlib~=3.8.3
seaborn~=0.13.2
loguru~=0.7.2
scikit-learn~=1.4.1.post1
pyfiglet~=0.8.post1
google-cloud-storage
keras_preprocessing
openpyxl
keras-cv
transformers
tf-keras
